requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.37.0
cryptography==41.0.5
Django==4.2.7
django-cloudinary-storage==0.3.0
django-embed-video==1.4.9
django-jsonform==2.20.2
django-paypal==2.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.2.0
paypalrestsdk==1.13.3
Pillow==10.1.0
pycparser==2.21
pyOpenSSL==23.3.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tmdbv3api==1.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
